More updates and examples, don't make volume for data dir
Making a volume for the data directory actually causes more issues than
it solves on Nomad - ultimately, it can't be shared properly with
the allocations that Nomad creates and ultimately breaks things like
shared secrets or dispatched jobs. Hence, it has been removed.

Also, the data directory that Nomad uses needs to be shared to the host
in most cases, so that docker can see the directory and that the
aforementioned secret and dispatched jobs directories can be seen
properly. This is also needs to be passed in somehow, so the
NOMAD_DATA_DIR and NOMAD_CONFIG_DIR variables have been exposed (the
second one mainly for brevity).

All of the content has been updated to reflect this, with the README
getting another update.