Shuffle build deps around again
This ensures that we can adjust the go and Alpine versions around
while ensuring that the build process remains stable.

Nomad still needs Node 8 to build, however we want to be able to use
a later version of Go when Nomad builds with it (possibly in Nomad
0.9.0). While Alpine 3.8.2 has Go 1.10.8 and Node 8, it probably
won't stay that way. Node is the heavier dependency to build, so we
have shifted to using the Node image instead of Go as the base, and
we just build the appropriate Go version bootstrapping off of the
distro-installed version.

The base image for the Nomad container has also been updated to 3.9,
the latest Alpine image. This should be fine barring any hypothetical
(and probably nonexistent) edge case caused by the minor version bump
to musl (1.1.19-r10 to 1.1.20-r3).